requirements.txt
# Dashboard & API
streamlit>=1.50.0
python-dotenv>=1.2.0
fastapi>=0.128.0
uvicorn>=0.39.0
pydantic>=2.0.0
APScheduler>=3.11.0

# Data ingestion & cleaning
google-play-scraper>=1.2.0
langdetect>=1.0.9
emoji>=2.15.0
presidio_analyzer>=2.2.0
presidio_anonymizer>=2.2.0
spacy>=3.8.0
phonenumbers>=9.0.0

# Spacy English model (required by presidio)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# LLM & clustering
groq>=1.0.0
sentence-transformers>=2.0.0
scikit-learn>=1.6.0
numpy>=2.0.0
torch>=2.0.0

# HTTP
requests>=2.32.0
httpx>=0.28.0
